+++ src/adapters-browser.ts
@@ -1,6 +1,4 @@
 export {
   BrowserAdapter,
-  WasmAdapter,
   type HlsDownloaderBrowserAdapter,
-  type HlsDownloaderWasmAdapter,
 } from '@hls-downloader/adapters/browser';