+++ src/index.ts
@@ -3,13 +3,9 @@ export { HlsDownloader, type GetAdditionalOptions } from '@hls-downloader/core';
 export { default } from '@hls-downloader/core';
 export {
   BrowserAdapter,
-  WasmAdapter,
   type HlsDownloaderBrowserAdapter,
-  type HlsDownloaderWasmAdapter,
 } from '@hls-downloader/adapters/browser';
 export {
   NodeAdapter,
-  RustAdapter,
   type HlsDownloaderNodeAdapter,
-  type HlsDownloaderRustAdapter,
 } from '@hls-downloader/adapters/node';